Splits CampaignList into container and item components
