+++ client/src/Components/CampaignList/CampaignListContainer.js
@@ -1,35 +1,15 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 import { Grid, PageHeader, Table } from 'react-bootstrap';
 import HttpRequest from '../HttpRequest';
+import CampaignListItem from './CampaignListItem';
 
 const helpers = require('../../helpers');
 
 export default class CampaignList extends React.Component {
-  static renderRow(campaign) {
-    const campaignId = campaign.id;
-    const triggers = campaign.topics[0].triggers;
-
-    return (
-      <tr key={campaignId}>
-        <td>{campaignId}</td>
-        <td>
-          <Link to={`campaigns/${campaignId}`}>
-            <strong>{campaign.title}</strong>
-          </Link>
-        </td>
-        <td>{triggers ? triggers.join(', ') : null}</td>
-        <td>{campaign.status}</td>
-      </tr>
-    );
-  }
-
   constructor(props) {
     super(props);
-
     this.requestPath = helpers.getCampaignsPath();
   }
-
   render() {
     return (
       <Grid>
@@ -44,7 +24,7 @@ export default class CampaignList extends React.Component {
                   <th>Triggers</th>
                   <th>Status</th>
                 </tr>
-                {res.map(campaign => CampaignList.renderRow(campaign))}
+                {res.map(campaign => <CampaignListItem key={campaign.id} campaign={campaign} />)}
               </tbody>
             </Table>)
           }
